refactor(install): subsystem-owned defaults and resolution chains, shared stream choreography, one byte formatter, and a four-file browser split
