+++ Sources/HedosKernel/Discovery/DiscoveryService.swift
@@ -35,22 +35,7 @@ public struct DiscoverySummary: Sendable {
         }
         let models = totalCount == 1 ? "1 model" : "\(totalCount) models"
         let breakdown = parts.isEmpty ? "" : " — \(parts.joined(separator: ", "))"
-        return "Found \(models) on this Mac\(breakdown). Total: \(Self.formatBytes(totalBytes))."
-    }
-
-    public static func formatBytes(_ bytes: Int64) -> String {
-        switch bytes {
-        case (1 << 30)...:
-            let value = Double(bytes) / Double(1 << 30)
-            let formatted = String(format: "%.1f", value)
-            return "\(formatted.hasSuffix(".0") ? String(formatted.dropLast(2)) : formatted) GB"
-        case (1 << 20)...:
-            return "\(bytes >> 20) MB"
-        case 1024...:
-            return "\(bytes >> 10) KB"
-        default:
-            return "\(bytes) B"
-        }
+        return "Found \(models) on this Mac\(breakdown). Total: \(ByteFormat.string(totalBytes))."
     }
 }
