+++ Sources/HedosKernel/Records/HardwareProfile.swift
@@ -3,15 +3,18 @@ import Foundation
 public struct HardwareProfile: Sendable, Hashable {
     public let chip: String
     public let ramGB: Int
+    public let totalMemoryBytes: UInt64
 
-    public init(chip: String, ramGB: Int) {
+    public init(chip: String, ramGB: Int, totalMemoryBytes: UInt64? = nil) {
         self.chip = chip
         self.ramGB = ramGB
+        self.totalMemoryBytes = totalMemoryBytes ?? UInt64(max(ramGB, 1)) << 30
     }
 
     public static let current = HardwareProfile(
         chip: readChip(),
-        ramGB: max(1, Int(ProcessInfo.processInfo.physicalMemory / (1 << 30))))
+        ramGB: max(1, Int(ProcessInfo.processInfo.physicalMemory / (1 << 30))),
+        totalMemoryBytes: ProcessInfo.processInfo.physicalMemory)
 
     public var summary: String {
         "\(chip) · \(ramGB) GB unified"
